backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
jinja2
python-dotenv
